Fixed wheel package build
Wheel package build is not working correctly.
