+++ po/en/LC_MESSAGES/PyPlatformGame.po
@@ -7,7 +7,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: PROJECT VERSION\n"
 "Report-Msgid-Bugs-To: EMAIL@ADDRESS\n"
-"POT-Creation-Date: 2022-06-14 12:49+0300\n"
+"POT-Creation-Date: 2022-06-14 22:52+0300\n"
 "PO-Revision-Date: 2022-06-14 01:45+0300\n"
 "Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
 "Language: en\n"
