+++ setup.cfg
@@ -1,0 +1,34 @@
+[metadata]
+name = PyPlatformGame
+version = 0.0.1
+url = https://github.com/antonnikvl/PyPlatformGame
+author = Anton N., Egor F., Michael D.
+author_email = antonnikvl@gmail.com
+description = Exam project for PythonDevelopment2022 course
+long_description = file: README.md
+classifiers =
+    Programming Language :: Python :: 3
+    Operating System :: OS Independent
+
+[options]
+packages = find_namespace:
+include_package_data = True
+install_requires =
+   pygame
+   pyglm
+   numpy
+   pyopengl
+   esper
+   pywavefront
+
+[options.entry_points]
+console_scripts =
+    PyPlatformGame = PyPlatformGame.main:main
+
+[options.package_data]
+PyPlatformGame =
+    */*/PyPlatformGame.mo
+    assets/*/*
+    assets/*
+    shaders/*
+
